Updated helper code for profiler attachment
